Transitions to CGFloat missed in the previous commit
git-svn-id: svn+ssh://svn.savannah.nongnu.org/gap/trunk@2280 d25b4930-3292-e311-b074-0010c6de5da8

+++ user-apps/Graphos/GRBox.m
@@ -337,7 +337,7 @@
     [endControlPoint moveToPoint: NSMakePoint(pos.x + size.width, pos.y + size.height)];
 }
 
-- (void)setZoomFactor:(float)f
+- (void)setZoomFactor:(CGFloat)f
 {
     [super setZoomFactor:f];
