+++ user-apps/Graphos/GRCircle.m
@@ -335,7 +335,7 @@
     [endControlPoint moveToPoint: NSMakePoint(pos.x + size.width, pos.y + size.height)];
 }
 
-- (void)setZoomFactor:(float)f
+- (void)setZoomFactor:(CGFloat)f
 {
     [super setZoomFactor:f];
